% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_launcher.R
\name{run_llm_linelist_app}
\alias{run_llm_linelist_app}
\title{Run the Packaged Shiny App}
\usage{
run_llm_linelist_app(display.mode = c("normal", "showcase"), ...)
}
\arguments{
\item{display.mode}{Shiny display mode.}

\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Invisibly returns the result of \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\description{
Launch the installed line list extraction Shiny application.
}
